fix: missing pull request validation checks
The pull request validation checks have two attributes to validate if
the checks are successful. The attributes are `conclusion` and `status`.
(https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#list-check-runs-for-a-git-reference)

This commit enhances the logic to check for the `conclusion` and `status`
of the checks together. And only merge the pull request if both are
successful.